update(now, all the data directories are stored in different locations to avoid conflict with other neovim configurations.)
First of all, this isn't complete. I need to create an installation
script that'll set everything up, but for now, I think this is the most
optimal solution because if we haulted on setuping this up, later on
down the road, it would've been a catastraphy to re-shape the entire
code structure.

- Packer plugin directory: ~/.local/share/nvcodark/site/pack/packer/{opt,start}
- Cache directory: ~/.cache/nvcodark
- Config directory: ~/.config/nvcodark

TODO: Create the installation script (Aug 14 2022 Sun (20:13:32), singularis)

Signature:

-------------------------------------------------------------------------------
Hashem A. Damrah <dermdorm2@gmail.com>

Mail:                     dermdorm2@gmail.com
Web:                      https://damrah.netlify.app
-------------------------------------------------------------------------------